+++ jpcd-yew/Cargo.toml
@@ -16,3 +16,4 @@ log = "0.4"
 wasm-logger = "0.2"
 gloo = "0.8"
 jpcd_game_of_life = "0.0.2"
+yew-router = "0.17.0"
