+++ jpcd-yew/conway.css
@@ -1,0 +1,44 @@
+.game-area {
+  width: 94%;
+  margin: 20px auto;
+}
+
+.game-container {
+  margin: 20px 0 0px 0;
+}
+
+.game-of-life {
+  display: inline-block;
+  background-color: aliceblue;
+  width: max-content;
+  overflow: hidden;
+}
+
+.game-buttons {
+  width: 100%;
+  margin-top: 20px;
+}
+
+.game-row {
+  line-height: 0;
+}
+
+.game-cell {
+  display: inline-block;
+  width: 8px;
+  height: 8px;
+  border: 1px solid #ccc;
+}
+
+.dead-cell {
+  background-color: white;
+}
+
+.alive-cell {
+  background-color: black;
+}
+
+.game-buttons {
+  width: 100%;
+  margin-top: 20px;
+}
