--- jpcd-yew/index.css
@@ -1,86 +1,0 @@
-html,
-body {
-  height: 100%;
-  margin: 0;
-}
-
-body {
-  align-items: center;
-  display: flex;
-  justify-content: center;
-
-  background: linear-gradient(to bottom right, #444444, #009a5b);
-  font-size: 1.5rem;
-}
-
-
-.app-header {
-  height: 130px;
-  color: aliceblue;
-}
-
-.app-title {
-  font-size: 24px;
-  width: 100%;
-  font-weight: 100;
-  text-align: center;
-  -webkit-text-rendering: optimizeLegibility;
-  -moz-text-rendering: optimizeLegibility;
-  text-rendering: optimizeLegibility;
-}
-
-.app-footer {
-  height: 30px;
-  padding: 10px;
-}
-
-.footer-text {
-  color: aliceblue;
-  padding-left: 20px;
-  font-size: 14px;
-}
-
-.game-area {
-  width: 94%;
-  margin: 20px auto;
-}
-
-.game-container {
-  margin: 20px 0 0px 0;
-}
-
-.game-of-life {
-  display: inline-block;
-  background-color: aliceblue;
-  width: max-content;
-  overflow: hidden;
-}
-
-.game-buttons {
-  width: 100%;
-  margin-top: 20px;
-}
-
-.game-row {
-  line-height: 0;
-}
-
-.game-cell {
-  display: inline-block;
-  width: 8px;
-  height: 8px;
-  border: 1px solid #ccc;
-}
-
-.dead-cell {
-  background-color: white;
-}
-
-.alive-cell {
-  background-color: black;
-}
-
-.game-buttons {
-  width: 100%;
-  margin-top: 20px;
-}
