+++ jpcd-yew/index.html
@@ -2,9 +2,14 @@
 <html>
   <head>
     <meta charset="utf-8" />
-    <title>Conway's game of life</title>
+    <title>JP Cloud Dev</title>
     <link data-trunk rel="rust" />
-    <link data-trunk rel="css" href="index.css" />
+    <link
+            rel="stylesheet"
+            href="https://cdn.jsdelivr.net/npm/bulma@0.9.0/css/bulma.min.css"
+    />
+    <link data-trunk rel="css" href="index.scss" />
+    <link data-trunk rel="css" href="conway.css" />
   </head>
   <body>
   </body>
